requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Flask==2.3.3
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
newsapi-python==0.2.7
requests==2.31.0
urllib3==2.0.4
Werkzeug==2.3.7
zipp==3.16.2
